requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.10.0
pyyaml>=6.0.1
httpx>=0.26.0
pytest>=8.0.0
